rename 'components' to 'ViewFunctions' since that is what they are :joy:
