+++ src/ViewFunctions/Counter.js
