+++ src/index.js
@@ -1,7 +1,7 @@
 import { app } from 'hyperapp';
 import actions from './actions';
 import state from './state';
-import view from './components/Counter';
+import view from './ViewFunctions/Counter';
 
 const dispatch = app({
   state,
